move getscrollbarwidth to helperfuncs
